src/Collaborative-Distillation/requirements.txt
scipy==1.4.1
torchvision==0.2.1
#torch==0.4.1
matplotlib==3.1.3
numpy==1.18.1
Pillow==8.2.0
